Refactoring of web API with Mongo sample app
