+++ aspnetcore/tutorials/first-mongo-app/sample/BookstoreAPI/appsettings.json
@@ -1,4 +1,7 @@
 {
+  "ConnectionStrings": {
+    "BookstoreDb": "mongodb://localhost:27017"
+  },
   "Logging": {
     "IncludeScopes": false,
     "Debug": {
